+++ test/unit/test_session.py
@@ -1150,3 +1150,27 @@ def test_plugin_arguments(mocker):
     )
     assert mock.called
     mock.assert_called_once_with([], [session], {"test": "value"})
+
+
+def test_close(session):
+    '''Close session.'''
+    assert session.closed is False
+    session.close()
+    assert session.closed is True
+
+
+def test_close_already_closed_session(session):
+    '''Close session that is already closed.'''
+    session.close()
+    assert session.closed is True
+    session.close()
+    assert session.closed is True
+
+
+def test_server_call_after_close(session):
+    '''Fail to issue calls to server after session closed.'''
+    session.close()
+    assert session.closed is True
+
+    with pytest.raises(ftrack_api.exception.ConnectionClosedError):
+        session.query('User').first()