+++ test/unit/test_session.py
@@ -1150,3 +1150,45 @@ def test_plugin_arguments(mocker):
     )
     assert mock.called
     mock.assert_called_once_with([], [session], {"test": "value"})
+
+
+def test_reset_attribute(session, user):
+    '''Reset user api key.'''
+    matching = session.get(
+        *ftrack_api.inspection.identity(user)
+    )
+
+    current_key = matching.get(
+        'api_key'
+    )
+
+    session.reset_attributes(
+        'User', matching.get('id'), ['api_key']
+    )
+
+    session.commit()
+
+    assert current_key != matching.get('api_key')
+
+
+@pytest.mark.parametrize('attribute', [
+    ('id',),
+    ('email',)
+
+], ids=[
+    'Fail resetting primary key',
+    'Fail resetting attribute without default value',
+])
+def test_fail_reset_attribute(session, user, attribute):
+    '''Fail trying to rest invalid attributes.'''
+
+    matching = session.get(
+        *ftrack_api.inspection.identity(user)
+    )
+
+    session.reset_attributes(
+        'User', matching.get('id'), attribute
+    )
+
+    with pytest.raises(ftrack_api.exception.ServerError):
+        session.commit()