+++ doc/release/release_notes.rst
@@ -14,6 +14,15 @@ Release Notes
     .. change:: new
         :tags: session
 
+        Support using a :class:`Session` as a context manager to aid closing of
+        session after use::
+
+            with ftrack_api.Session() as session:
+                # Perform operations with session.
+
+    .. change:: new
+        :tags: session
+
         :meth:`Session.close` automatically called on Python exit if session not
         already closed.